% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggr_Attrs.R
\name{Aggr_Attrs}
\alias{Aggr_Attrs}
\title{Aggregate Attributes}
\usage{
Aggr_Attrs(tree, attrs = NULL, aggFun = sum)
}
\arguments{
\item{tree}{data.tree}

\item{attrs}{character, optional. Vector of attributes to aggregate. Default
is all attributes.}

\item{aggFun}{function, optional. The function to be used to aggregate each
attribute. Default is \code{sum}.}
}
\value{
data.tree
}
\description{
Aggregate Attributes
}
