% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggr_Scaled.R
\name{Aggr_Scaled}
\alias{Aggr_Scaled}
\title{Aggregate Scaled Values}
\usage{
Aggr_Scaled(x)
}
\arguments{
\item{x}{numeric, scaled values to be aggregated}
}
\description{
Aggregates values that have been scaled by averaging them and then
multiplying by the square root of n
}
