% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggr_Attrs.R
\name{.Aggr_Attr}
\alias{.Aggr_Attr}
\title{Aggregate Attribute}
\usage{
.Aggr_Attr(attr, aggFun = sum)
}
\arguments{
\item{attr}{character, name of the attribute to be aggregated up the tree}

\item{aggFun}{function, the function to be used for aggregation. Function
should take a numeric vector as its only argument and return a numeric singleton}
}
\description{
Aggregate Attribute
}
