% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testing_Apperati.R
\name{.Make_Test_DF}
\alias{.Make_Test_DF}
\title{Make Test Data.Frame}
\usage{
.Make_Test_DF()
}
\value{
data.frame
}
\description{
Internal function that constructs a test data.frame for testing purposes
}
