% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrame_Functions.R
\name{is.Node}
\alias{is.Node}
\title{Is object a data.tree Node}
\usage{
is.Node(x)
}
\arguments{
\item{x}{object. If \code{x} is a list of length greater than 1, each element
of \code{x} will be checked individually. Otherwise, only \code{x} will be
checked.}
}
\value{
logical
}
\description{
Is object a data.tree Node
}
