% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isolate_Aggr.R
\name{Isolate_Aggr}
\alias{Isolate_Aggr}
\title{Isolate and Aggregate}
\usage{
Isolate_Aggr(df, isolate, over, attrs = NULL, aggFun = sum)
}
\arguments{
\item{df}{data.frame}

\item{isolate}{symbol, a pathString (or other) field which will be held
constant}

\item{over}{symbol, a pathString field which will be aggregated according to
hierarchy}

\item{attrs}{character, optional. Vector of attributes to aggregate. Default
is all attributes.}

\item{aggFun}{function, optional. The function to be used to aggregate each
attribute. Default is \code{sum}.}
}
\value{
data.frame
}
\description{
This function is used to make a list of trees, one for each unique value in
the isolate column in the data.frame \code{df}
}
