% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isolate_Aggr.R
\name{Make_Isolate_Trees}
\alias{Make_Isolate_Trees}
\title{Make Isolate Trees}
\usage{
Make_Isolate_Trees(df, isolate, over)
}
\arguments{
\item{df}{data.frame}

\item{isolate}{symbol, a column in \code{df} which will be isolated}

\item{over}{symbol, a column in \code{df} which will be aggregated over}
}
\value{
list of data.trees
}
\description{
This function is used to make a list of trees, one for each unique value in
the isolate column in the data.frame \code{df}
}
