% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print_Tree.R
\name{Print_Tree}
\alias{Print_Tree}
\title{Print Tree}
\usage{
Print_Tree(tree, attrs = NULL)
}
\arguments{
\item{tree}{data.tree}

\item{attrs}{character, vector of node attributes whose values are to be
printed}
}
\value{
data.tree (invisibly)
}
\description{
Print Tree
}
