% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrame_Functions.R
\name{transpose_tree_df}
\alias{transpose_tree_df}
\title{Transpose Tree Data.Frame}
\usage{
transpose_tree_df(df)
}
\arguments{
\item{df}{data.frame, should contain a column named 'pathString'}
}
\value{
data.frame
}
\description{
Transposes a data.frame such that the values in the column named 'pathString'
in the original data.frame become the new column names, and the column names
in the original data.frame become the new values in the column named
'pathString'.
}
\details{
Note that if the attribute columns are not all of the same type, they WILL be
coerced to the same type using the usual order of preference.
(logical to integer to numeric to character)
}
