% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrame_Functions.R
\name{tree_to_df}
\alias{tree_to_df}
\title{To Data.Frame All}
\usage{
tree_to_df(tree, attrs = NULL, minimal = TRUE)
}
\arguments{
\item{tree}{data.tree}

\item{attrs}{character, optional}
}
\value{
data.frame
}
\description{
Function to convert data.tree object back to a data.frame, with all attrs.
By using ToDataFrameTable, you would get only the leaf-level information.
ToDataFrameTree gives aggregate values calculated at the different levels of
the hierarchy as well, which is needed to compute values for metrics at each
of the levels.
}
